+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added index.php that load the autoloader and the bootstrapping
 - Added bootstrapping to bioman CLI to get access to the global functions
 - Added facade Path with function for joining paths
+- Added global function `printLn`
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2